+++ src/utils/internal/index.ts
@@ -7,4 +7,6 @@
 
 export * from "./errorHandler.js";
 export * from "./logger.js";
+export * from "./performance.js";
 export * from "./requestContext.js";
+